+++ src/components/UserPage/PetsData/PetsData.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import PetsList from './PetsList';
+import PetsList from '../PetsList/PetsList';
 function PetsData() {
   return (
     <>
